+++ src/main/java/com/gravity9/jsonpatch/ReplaceOperation.java
@@ -45,10 +45,6 @@ public final class ReplaceOperation extends PathValueOperation {
 	public JsonNode applyInternal(final JsonNode node) throws JsonPatchException {
 		final String jsonPath = JsonPathParser.tmfStringToJsonPath(path);
 		final DocumentContext nodeContext = JsonPath.parse(node.deepCopy());
-		final JsonNode nodeAtPath = nodeContext.read(jsonPath);
-		if (nodeAtPath == null) {
-			throw new JsonPatchException(BUNDLE.getMessage("jsonPatch.noSuchPath"));
-		}
 		final JsonNode replacement = value.deepCopy();
 		if (path.isEmpty()) {
 			return replacement;