+++ src/main/java/com/gravity9/jsonpatch/TestOperation.java
@@ -51,11 +51,8 @@ public final class TestOperation extends PathValueOperation {
 	public JsonNode applyInternal(final JsonNode node) throws JsonPatchException {
 		final String jsonPath = JsonPathParser.tmfStringToJsonPath(path);
 		final JsonNode tested = JsonPath.parse(node.deepCopy()).read(jsonPath);
-		if (tested == null) {
-			throw new JsonPatchException(BUNDLE.getMessage("jsonPatch.noSuchPath"));
-		}
 		if (!EQUIVALENCE.equivalent(tested, value)) {
-			throw new JsonPatchException(BUNDLE.getMessage("jsonPatch.valueTestFailure"));
+			throw JsonPatchException.valueTestFailure(value, tested);
 		}
 		return node.deepCopy();
 	}